requirements.txt
Flask==3.0.0
PyYAML==5.4.1
requests==2.31.0